backend/requirements.txt
google-api-python-client==2.128.0
google-cloud-logging==3.10.0
google-cloud-firestore==2.16.0
firebase-admin==6.3.0
Flask==2.3.3
Flask_Cors==4.0.1
